Subida printf realizado

+++ Cursus/C/printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: adriescr <adriescr@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/04/24 18:50:02 by adriescr          #+#    #+#             */
-/*   Updated: 2025/04/25 23:40:48 by adriescr         ###   ########.fr       */
+/*   Updated: 2025/04/27 00:38:11 by adriescr         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int			ft_putchar(char c);
 int			ft_putstr(char *str);
 int			ft_putnbr(int n);
 int			ft_putnbr_unsigned(unsigned int n);
-int			ft_putnbr_base(unsigned int nbr, int upper);
+int			ft_putnbr_base(unsigned long nbr, int upper);
 int			ft_putptr(void *ptr);
 
 #endif
